BUild pynacl with included libsodium source code
Key factor was to build cffi first and then pynacl because we are
using no-build-isolation flag. According to pip documentation:
https://pip.pypa.io/en/stable/reference/pip/
Users supplying this flag are responsible for ensuring the build
environment is managed appropriately (including ensuring that all
required build dependencies are installed).
Also see the use of verboseness, builddirectory, clean and log flags
for pip